refactor(general): refactor code
