+++ ozpcenter/api/contact_type/serializers.py
@@ -42,3 +42,27 @@ class ListingContactTypeSerializer(serializers.ModelSerializer):
             ret['name'] = access_control_instance.anonymize_value('contact_type_name')
 
         return ret
+
+
+class ContactSerializer(serializers.ModelSerializer):
+    contact_type = ListingContactTypeSerializer()
+
+    class Meta:
+        model = models.Contact
+        fields = '__all__'
+
+    def to_representation(self, data):
+        access_control_instance = plugin_manager.get_system_access_control_plugin()
+        ret = super(ContactSerializer, self).to_representation(data)
+
+        # Used to anonymize usernames
+        anonymize_identifiable_data = system_anonymize_identifiable_data(self.context['request'].user.username)
+
+        if anonymize_identifiable_data:
+            ret['secure_phone'] = access_control_instance.anonymize_value('secure_phone')
+            ret['unsecure_phone'] = access_control_instance.anonymize_value('unsecure_phone')
+            ret['secure_phone'] = access_control_instance.anonymize_value('secure_phone')
+            ret['name'] = access_control_instance.anonymize_value('name')
+            ret['organization'] = access_control_instance.anonymize_value('organization')
+            ret['email'] = access_control_instance.anonymize_value('email')
+        return ret
