+++ ozpcenter/api/contact_type/views.py
@@ -60,3 +60,51 @@ class ContactTypeViewSet(viewsets.ModelViewSet):
     queryset = model_access.get_all_contact_types()
     serializer_class = serializers.ContactTypeSerializer
     permission_classes = (permissions.IsAppsMallStewardOrReadOnly,)
+
+
+class ContactViewSet(viewsets.ModelViewSet):
+    """
+    ModelViewSet for getting all contacts for a given listing
+
+    Access Control
+    ===============
+    - All users can view
+
+    URIs
+    ======
+    GET /api/contact
+    Summary:
+        Get a list of all system-wide Contact entries
+    Response:
+        200 - Successful operation - [ContactSerializer]
+
+    POST /api/contact/
+    Summary:
+        Add a Contact
+    Request:
+        data: ContactSerializer Schema
+    Response:
+        200 - Successful operation - ContactSerializer
+
+    GET /api/contact/{pk}
+    Summary:
+        Find a Contact Entry by ID
+    Response:
+        200 - Successful operation - ContactSerializer
+
+    PUT /api/contact/{pk}
+    Summary:
+        Update a Contact Entry by ID
+
+    PATCH /api/contact/{pk}
+    Summary:
+        Update (Partial) a Contact Entry by ID
+
+    DELETE /api/contact/{pk}
+    Summary:
+        Delete a Contact Entry by ID
+    """
+    # TODO: validate permission_classes
+    permission_classes = (permissions.IsUser,)
+    queryset = model_access.get_all_contacts()
+    serializer_class = serializers.ContactSerializer
