+++ ozpcenter/api/listing/model_access_es.py
@@ -11,11 +11,11 @@ from django.conf import settings
 from django.http.request import QueryDict
 from rest_framework import serializers
 
-from ozpcenter import models
-from ozpcenter import constants
-from plugins.plugin_manager import system_has_access_control
-from ozpcenter.api.listing import elasticsearch_util
 from ozpcenter.api.listing.elasticsearch_util import elasticsearch_factory
+from ozpcenter.api.listing import elasticsearch_util
+from ozpcenter import constants
+from ozpcenter import models
+from plugins.plugin_manager import system_has_access_control
 
 
 logger = logging.getLogger('ozp-center.' + str(__name__))
