+++ ozpcenter/api/listing/urls.py
@@ -34,7 +34,6 @@ nested_router.register(r'rejection', views.ListingRejectionViewSet, base_name='r
 router.register(r'listingtype', views.ListingTypeViewSet)
 router.register(r'screenshot', views.ScreenshotViewSet)
 router.register(r'tag', views.TagViewSet)
-router.register(r'contact', views.ContactViewSet)
 
 # Wire up our API using automatic URL routing.
 urlpatterns = [
